Separate default styles from deck-styles
