+++ packages/viz/src/index.ts
@@ -5,12 +5,19 @@ export { DOSource } from './lib/sources';
 // Style helpers
 export {
   basicStyle,
+  BasicOptionsStyle,
   colorBinsStyle,
+  ColorBinsOptionsStyle,
   colorCategoriesStyle,
+  ColorCategoriesOptionsStyle,
   colorContinuousStyle,
+  ColorContinuousOptionsStyle,
   sizeBinsStyle,
+  SizeBinsOptionsStyle,
   sizeCategoriesStyle,
-  sizeContinuousStyle
+  SizeCategoriesOptionsStyle,
+  sizeContinuousStyle,
+  SizeContinuousOptionsStyle
 } from './lib/style';
 
 // Basemap helpers
