+++ packages/viz/src/lib/style/helpers/index.ts
@@ -1,7 +1,19 @@
 export { basicStyle } from './basic-style';
-export { colorBinsStyle } from './color-bins-style';
-export { colorCategoriesStyle } from './color-categories-style';
-export { colorContinuousStyle } from './color-continuous-style';
-export { sizeBinsStyle } from './size-bins-style';
-export { sizeCategoriesStyle } from './size-categories-style';
-export { sizeContinuousStyle } from './size-continuous-style';
+export { colorBinsStyle, ColorBinsOptionsStyle } from './color-bins-style';
+export {
+  colorCategoriesStyle,
+  ColorCategoriesOptionsStyle
+} from './color-categories-style';
+export {
+  colorContinuousStyle,
+  ColorContinuousOptionsStyle
+} from './color-continuous-style';
+export { sizeBinsStyle, SizeBinsOptionsStyle } from './size-bins-style';
+export {
+  sizeCategoriesStyle,
+  SizeCategoriesOptionsStyle
+} from './size-categories-style';
+export {
+  sizeContinuousStyle,
+  SizeContinuousOptionsStyle
+} from './size-continuous-style';
