+++ packages/viz/src/lib/style/index.ts
@@ -1,13 +1,21 @@
-export * from './default-styles';
+export { defaultStyles } from './default-styles';
 
 export { Style, StyleProperties } from './Style';
 
 export {
   basicStyle,
   colorBinsStyle,
+  ColorBinsOptionsStyle,
   colorCategoriesStyle,
+  ColorCategoriesOptionsStyle,
   colorContinuousStyle,
+  ColorContinuousOptionsStyle,
   sizeBinsStyle,
+  SizeBinsOptionsStyle,
   sizeCategoriesStyle,
-  sizeContinuousStyle
+  SizeCategoriesOptionsStyle,
+  sizeContinuousStyle,
+  SizeContinuousOptionsStyle
 } from './helpers';
+
+export { getStyles, getStyleValue, BasicOptionsStyle } from './deck-styles';
